Get logging off the ground
It is not efficient at the moment. Right now, all formatting is done at the call site, and dynamic memory is involved. Ideally, I'd put some reasonable limits on string length and # of format arguments so that I could move all formatting onto a background thread via a waitfree queue of log records.

Also, everything dumps to stdout right now. I'd love to be able to tell the logger to push to different sinks. Then, in some sort of CLI monitor, I could set up a sink that queues up log messages to be displayed to an interactive user.

But this is okay :)