Add some loops for watching robot values
Right now, we can log them to see what we're working with and what values we can and can't hardcode when working out our drive logic.

It would be really cool if we had a custom visualizer (written with opengl or something) that showed the faces recognized and with some graphs for these monitored values.

But this is cool, too :)